Change int runtime type from i32 to i64
Update the Protosept runtime to use i64 for the `int` type, matching
the language specification. Changes across bytecode (Ldi instruction),
interpreter (Data::Int, Exception, comparison/logical ops, binary_op_int),
builtin functions (string/array length casts), codegen (literal and enum
variant index emissions), and test harness (expected value parsing).

+++ p7-cli/src/test_harness.rs
@@ -189,7 +189,7 @@ fn run_test_case(
 
     let is_match = match &p7_result {
         P7Value::Int(actual_val) => expected_value_str
-            .parse::<i32>()
+            .parse::<i64>()
             .map_or(false, |expected_val| *actual_val == expected_val),
         P7Value::Float(actual_val) => expected_value_str
             .parse::<f64>()
